+++ src/views/calculatorIndex.js
@@ -1,0 +1,7 @@
+export default function CalculatorIndex() {
+  return (
+    <div>
+      <h1>Calculatrice</h1>
+    </div>
+  )
+}
