+++ src/views/home.js
@@ -1,0 +1,7 @@
+export default function Home() {
+  return (
+    <div>
+      <h1>Dashboard</h1>
+    </div>
+  )
+}
